requirements.txt
pydantic-ai-slim[openai,logfire,groq]==1.87.0
logfire[fastapi]==4.32.1
fastapi==0.136.1
uvicorn==0.35.0
httpx==0.28.1
python-dotenv==1.1.0
tavily-python==0.3.6
upstash-redis==1.3.0
